backend/requirements.txt
amqp==5.3.1
asgiref==3.6.0
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
Django==4.2.5
django-celery-beat==2.8.1
django-ckeditor==6.7.3
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-recaptcha==4.1.0
django-timezone-field==7.1
djangorestframework==3.16.0
gunicorn==21.2.0
idna==3.10
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyMySQL==1.1.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
requests==2.32.4
six==1.17.0
sqlparse==0.4.4
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
